+++ database/migrations/2021_03_25_072244_add_brands_field.php
@@ -13,14 +13,16 @@ class AddBrandsField extends Migration
      */
     public function up()
     {
-        Schema::table('products', function ($table) {
+        if (!Schema::hasColumn('products', 'brand_id')) {
+            Schema::table('products', function (Blueprint $table) {
             $table->bigInteger('brand_id')->unsigned()->nullable();
             $table->bigInteger('tax_category_id')->unsigned()->nullable();
             $table->foreign('brand_id')->references('id')->on('brands')->onDelete('set null');
             $table->foreign('tax_category_id')->references('id')->on('tax_categories')->onDelete('set null');
-        });
-    }
+                    });
+        }
 
+        }
     /**
      * Reverse the migrations.
      *
@@ -28,9 +30,6 @@ class AddBrandsField extends Migration
      */
     public function down()
     {
-        Schema::table('users', function ($table) {
-            $table->dropColumn('brand_id');
-            $table->dropColumn('tax_category_id');
-        });
+        // Reverse migration if needed
     }
 }
