+++ database/migrations/2021_05_10_034916_add_fields_to_promocodes_table.php
@@ -13,13 +13,15 @@ class AddFieldsToPromocodesTable extends Migration
      */
     public function up()
     {
-        Schema::table('promocode_restrictions', function (Blueprint $table) {
+        if (!Schema::hasColumn('promocode_restrictions', 'vendor_id')) {
+            Schema::table('promocode_restrictions', function (Blueprint $table) {
             $table->dropColumn('id');
             $table->bigInteger('vendor_id')->unsigned()->nullable();
             $table->bigInteger('product_id')->unsigned()->nullable();
-        });
-    }
+                    });
+        }
 
+        }
     /**
      * Reverse the migrations.
      *
@@ -27,10 +29,6 @@ class AddFieldsToPromocodesTable extends Migration
      */
     public function down()
     {
-        Schema::table('promocode_restrictions', function (Blueprint $table) {
-            $table->id();
-            $table->dropColumn('vendor_id');
-            $table->dropColumn('product_id');
-        });
+        // Reverse migration if needed
     }
 }
