+++ database/migrations/2021_05_21_103918_alter_order_products_for_rename_table_table.php
@@ -13,18 +13,13 @@ class AlterOrderProductsForRenameTableTable extends Migration
      */
     public function up()
     {
-        Schema::table('orders', function (Blueprint $table) {
+        if (!Schema::hasColumn('orders', 'coupon_id')) {
+            Schema::table('orders', function (Blueprint $table) {
             $table->dropColumn('status');
-        });
-        Schema::table('order_vendors', function (Blueprint $table) {
-            $table->renameColumn('dipatcher_fee', 'delivery_fee');
-            $table->bigInteger('coupon_id')->unsigned()->nullable();
-        });
-        Schema::table('order_products', function (Blueprint $table) {
-            DB::statement('ALTER TABLE `order_products` RENAME TO `order_vendor_products`');
-        });
-    }
+                    });
+        }
 
+        }
     /**
      * Reverse the migrations.
      *
@@ -32,7 +27,6 @@ class AlterOrderProductsForRenameTableTable extends Migration
      */
     public function down()
     {
-        Schema::table('order_products', function (Blueprint $table) {
-        });
+        // Reverse migration if needed
     }
 }
