+++ database/migrations/2021_06_16_052121_add_vendor_id_to_order_status.php
@@ -13,12 +13,14 @@ class AddVendorIdToOrderStatus extends Migration
      */
     public function up()
     {
-        Schema::table('vendor_order_statuses', function (Blueprint $table) {
+        if (!Schema::hasColumn('vendor_order_statuses', 'vendor_id')) {
+            Schema::table('vendor_order_statuses', function (Blueprint $table) {
             $table->bigInteger('vendor_id')->unsigned()->nullable();
             $table->foreign('vendor_id')->references('id')->on('vendors')->onDelete('cascade');
-        });
-    }
+                    });
+        }
 
+        }
     /**
      * Reverse the migrations.
      *
@@ -26,8 +28,6 @@ class AddVendorIdToOrderStatus extends Migration
      */
     public function down()
     {
-        Schema::table('vendor_order_statuses', function (Blueprint $table) {
-            $table->dropColumn('vendor_id');
-        });
+        // Reverse migration if needed
     }
 }
