+++ database/migrations/2021_07_08_064807_add_vendor_id_to_product_inquiries_table.php
@@ -13,13 +13,15 @@ class AddVendorIdToProductInquiriesTable extends Migration
      */
     public function up()
     {
-        Schema::table('product_inquiries', function (Blueprint $table) {
+        if (!Schema::hasColumn('product_inquiries', 'vendor_id')) {
+            Schema::table('product_inquiries', function (Blueprint $table) {
             $table->bigInteger('vendor_id')->unsigned()->nullable();
             $table->foreign('vendor_id')->references('id')->on('vendors')->onDelete('cascade');
             $table->bigInteger('product_variant_id')->unsigned()->nullable();
-        });
-    }
+                    });
+        }
 
+        }
     /**
      * Reverse the migrations.
      *
@@ -27,8 +29,6 @@ class AddVendorIdToProductInquiriesTable extends Migration
      */
     public function down()
     {
-        Schema::table('product_inquiries', function (Blueprint $table) {
-            //
-        });
+        // Reverse migration if needed
     }
 }
